requirements.txt
absl-py==0.15.0
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bio==1.3.3
biopython==1.79
biothings-client==0.2.6
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
debugpy==1.5.1
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml==0.7.1
Distance==0.1.3
docopt==0.6.2
entrypoints==0.3
Flask==2.2.3
flatbuffers==2.0
fonttools==4.32.0
gast==0.4.0
gevent==22.10.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.41.1
h5py==3.5.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.4.0
ipykernel==6.5.0
ipython @ file:///tmp/build/80754af9/ipython_1635943998418/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1611333125342/work
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==12.0.0
lxml==4.9.2
Markdown==3.3.4
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mygene==3.2.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
ncbi-genome-download==0.3.1
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
premailer==3.10.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
SciencePlots==1.0.9
scikit-learn==1.1.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
torch==1.10.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
utils==1.0.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.5.2
wrapt==1.13.3
yagmail==0.15.293
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
